+++ src/ops/agg_engine.h
@@ -17,19 +17,23 @@ bool agg_v2_can_handle(ray_graph_t* g, ray_op_t* op, ray_t* tbl);
 /* Precondition: agg_v2_can_handle(g, op, tbl) returned true. */
 ray_t* exec_group_v2(ray_graph_t* g, ray_op_t* op, ray_t* tbl);
 
-/* Dense group assignment over 1..16 key columns, first-occurrence order. */
+/* Dense group assignment over the group's key columns, first-occurrence
+ * order (key count unbounded since cut 3; the dense planner self-limits). */
 typedef struct {
     uint32_t* gids;       /* len = nrows */
     int64_t*  first_row;  /* len = ngroups; row index where each group first appeared */
     int64_t   ngroups;
 } agg_groups_t;
 
-/* Multi-key grouping (1..16 keys). Reads each key as an int64 (intern id for
- * SYM) and hashes the tuple. Assigns gids incrementally on first sight → gid
+/* Multi-key grouping, key count unbounded. Reads each key as an int64
+ * (intern id for SYM) and hashes the tuple. Assigns gids incrementally on first sight → gid
  * order == first-occurrence order; first_row[gid] records the row where the
  * group first appeared. Returns 0 on success (caller releases out via
- * agg_groups_free()), -1 on allocation failure. */
-int agg_group_keys(ray_t** key_cols, uint8_t n_keys, int64_t nrows, agg_groups_t* out);
+ * agg_groups_free()), -1 on allocation failure.
+ * n_keys is uint32_t: unbounded key count.  Cut-3 lifted both admission gates
+ * (the GROUP path and the keys-only DISTINCT path) and the fixed data[16]
+ * inside became an exact carve, so any key count groups correctly. */
+int agg_group_keys(ray_t** key_cols, uint32_t n_keys, int64_t nrows, agg_groups_t* out);
 
 /* Release the buffers an agg_groups_t holds (buddy-backed, NOT libc malloc — so
  * callers must use this, not free()).  Idempotent; NULLs the pointers. */
@@ -41,9 +45,11 @@ void agg_groups_free(agg_groups_t* out);
  * interning).  See agg_engine.c.  Precondition (caller-gated): keys are int/SYM
  * and every tbl column is fixed-width/SYM/STR/LIST.  Caller owns the table.
  * keep_syms (or NULL) lists the column syms a consumer references — non-key
- * columns NOT in it are dropped (projection pushdown); keys are always kept. */
+ * columns NOT in it are dropped (projection pushdown); keys are always kept.
+ * nk is uint32_t: unbounded key count (cut-3 lifted query.c's distinct gate;
+ * agg_group_keys carves its key table, so any nk groups correctly). */
 ray_t* agg_select_distinct(ray_t* tbl, ray_t** key_cols, const int64_t* key_syms,
-                           uint8_t nk, int64_t nrows,
+                           uint32_t nk, int64_t nrows,
                            const int64_t* keep_syms, int keep_n);
 
 /* Build a dense SoA per-group state array for one aggregate (vt), run a single
@@ -62,10 +68,10 @@ ray_t* agg_run_one(const agg_vtable_t* vt, ray_t* val_col,
 
 typedef struct {
     bool     ok;
-    uint8_t  n_keys;
-    int64_t  mins[16];      /* per-key min */
-    int64_t  ranges[16];    /* per-key (max-min+1) */
-    int64_t  strides[16];   /* composite packing: slot = sum_k (key_k - min_k)*strides[k] */
+    uint32_t n_keys;        /* mirrors ext->n_keys' width; value stays 1..16 (dense self-limit) */
+    int64_t  mins[16];      /* [16]: dense direct-index routing self-limits to <=16 keys (agg_dense_plan) */
+    int64_t  ranges[16];    /* [16]: dense direct-index routing self-limits to <=16 keys (agg_dense_plan) */
+    int64_t  strides[16];   /* [16]: dense self-limit <=16; composite packing: slot = sum_k (key_k - min_k)*strides[k] */
     int64_t  total_slots;   /* product of ranges */
 } dense_plan_t;
 
@@ -73,9 +79,14 @@ typedef struct {
  *  - every key type in {I64,I32,I16,U8,BOOL,DATE,TIME,TIMESTAMP,SYM} and NOT HAS_NULLS
  *  - every agg vtable is ACC_STREAMING (no buffered median/top)
  *  - product of per-key ranges <= DENSE_MAX_SLOTS (no overflow)
- * Does one min/max prescan over the key columns.  Sets out->ok accordingly. */
-bool agg_dense_plan(ray_t** key_cols, uint8_t n_keys,
-                    const agg_vtable_t** vts, uint8_t n_aggs,
+ * Does one min/max prescan over the key columns.  Sets out->ok accordingly.
+ * n_keys is uint32_t (untruncated ext->n_keys): dense direct-index routing
+ * self-limits to <=16 keys here (wider shapes are rejected to v2's unbounded
+ * hash/radix path), so the fixed [16] mins/ranges/strides are only ever read
+ * at <=16.  n_aggs is unused by dense eligibility (kept uint32_t so a widened
+ * ext->n_aggs never narrows at this call). */
+bool agg_dense_plan(ray_t** key_cols, uint32_t n_keys,
+                    const agg_vtable_t** vts, uint32_t n_aggs,
                     int64_t nrows, dense_plan_t* out);
 
 /* Result column name for a plain-column-input aggregate: input column name
